+++ CHANGELOG.md
@@ -3,7 +3,7 @@ CHANGELOG
 
 ## HEAD (Unreleased)
 * Update pulumi-terraform to fdafc00b3f
-
+* Out-of-the-box Azure Cloud Shell support ([#74](https://github.com/pulumi/pulumi-azure/issues/74))
 ___
 
 ## 0.19.3 (2019-07-22)